Improve the api about ledger.
